+++ src/libkern/kstring/strcmp.c
@@ -45,5 +45,5 @@ int strcmp(const char * str1, const char * str2)
         str2++;
     }
 
-    return (*(unsigned char *)str1) - (*(unsigned char *)str2);
+    return ((int)(*(unsigned char *)str1) - (int)(*(unsigned char *)str2));
 }
