+++ src/libkern/kstring/strncmp.c
@@ -46,5 +46,5 @@ int strncmp(const char * str1, const char * str2, size_t n)
         str2++;
     }
 
-    return (*(unsigned char *)str1) - (*(unsigned char *)str2);
+    return ((int)(*(unsigned char *)str1) - (int)(*(unsigned char *)str2));
 }
